+++ sources/ast/expansion/expr_bracket.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   expr_bracket.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dwisoky <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: dwisoky <dwisoky@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/01 20:01:02 by dwisoky           #+#    #+#             */
-/*   Updated: 2019/10/01 20:09:38 by dwisoky          ###   ########.fr       */
+/*   Updated: 2019/10/11 19:21:47 by dwisoky          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "arifmetic.h"
 
-size_t		expr_bracket(void)
+long		expr_bracket(void)
 {
-	size_t	return_value;
+	long	return_value;
 
 	if (g_lex_arif && g_lex_arif->type == LBRACKET_ARIF)
 	{
