+++ sources/ast/expansion/expr_equal.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   expr_equal.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dwisoky <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: dwisoky <dwisoky@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/29 19:09:41 by dwisoky           #+#    #+#             */
-/*   Updated: 2019/10/01 17:44:03 by dwisoky          ###   ########.fr       */
+/*   Updated: 2019/10/11 18:52:37 by dwisoky          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "arifmetic.h"
 
-size_t		expr_equal(void)
+long		expr_equal(void)
 {
-	size_t	right_value;
-	size_t	left_value;
+	long	right_value;
+	long	left_value;
 	int		type;
 
 	left_value = expr_relat_op();
