+++ sources/ast/expansion/expr_logic_and.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   expr_logic_and.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dwisoky <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: dwisoky <dwisoky@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/29 19:06:33 by dwisoky           #+#    #+#             */
-/*   Updated: 2019/10/01 17:50:46 by dwisoky          ###   ########.fr       */
+/*   Updated: 2019/10/11 18:49:53 by dwisoky          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "arifmetic.h"
 
-size_t		expr_logic_and(void)
+long		expr_logic_and(void)
 {
-	size_t	left_value;
-	size_t	right_value;
+	long	left_value;
+	long	right_value;
 
 	left_value = expr_equal();
 	if (g_error_arifmetic)
