+++ sources/ast/expansion/expr_relat_op.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   expr_relat_op.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dwisoky <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: dwisoky <dwisoky@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/29 19:26:44 by dwisoky           #+#    #+#             */
-/*   Updated: 2019/10/01 18:23:50 by dwisoky          ###   ########.fr       */
+/*   Updated: 2019/10/11 18:53:26 by dwisoky          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "arifmetic.h"
 
-size_t		expr_relat_op(void)
+long		expr_relat_op(void)
 {
-	size_t	left_value;
-	size_t	right_value;
+	long	left_value;
+	long	right_value;
 	int		type;
 
 	left_value = expr_shift();
